JavaCustomToolsFromLorenWang:::修复文件内容存储拼接后面存储无法存储问题

+++ JavaCustomToolsFromLorenWang/src/javabase/lorenwang/tools/file/JtlwFileOptionUtils.java
@@ -221,7 +221,9 @@ public class JtlwFileOptionUtils {
      */
     public Boolean writeToFile(File file, InputStream inputStream, boolean append) {
         //删除文件
-        deleteFile(file.getAbsolutePath());
+        if(!append) {
+            deleteFile(file.getAbsolutePath());
+        }
         //创建父级文件夹
         createDirectory(file.getAbsolutePath(), true);
         FileOutputStream fos = null;
